# Improve performance by bypassing costly style/entity processing for text-only blocks