@loicteixeira 2018-03-06
Sorry I didn't have time to review it before you merged it. It looks good though 👌